+++ UnityProject/Assets/Scripts/Managers/GameManager.cs
@@ -417,8 +417,6 @@ public partial class GameManager : MonoBehaviour, IInitialise
 			// Tell all clients that the countdown has finished
 			UpdateCountdownMessage.Send(true, 0);
 
-			CurrentRoundState = RoundState.Started;
-			EventManager.Broadcast(EVENT.RoundStarted);
 		}
 	}