+++ UnityProject/Assets/Scripts/Tilemaps/Behaviours/Objects/RegisterDoor.cs
@@ -61,7 +61,7 @@
 		}
 
 
-		public override bool IsPassableTo(Vector3Int leavingTo, bool isServer, GameObject context = null)
+		public override bool IsPassableFromInside(Vector3Int leavingTo, bool isServer, GameObject context = null)
 		{
 			if (isClosed && OneDirectionRestricted)
 			{
@@ -108,10 +108,10 @@
 			return true;
 		}
 
-		public override bool IsPassable( Vector3Int from, bool isServer, GameObject context = null)
+		public override bool IsPassableFromOutside( Vector3Int from, bool isServer, GameObject context = null)
 		{
 			// Entering and leaving is the same check
-			return IsPassableTo( from, isServer );
+			return IsPassableFromInside( from, isServer );
 		}
 
 		public override bool IsPassable(bool isServer, GameObject context = null)